Refactors Order and OrderEvent
Inherits from Lean object and create `AlphaStreamOrder` and `AlphaStreamOrderEvent`
