+++ QuantConnect.AlphaStream/Requests/GetAlphaOrdersRequest.cs
@@ -1,7 +1,7 @@
-﻿using System.Collections.Generic;
-using QuantConnect.AlphaStream.Infrastructure;
+﻿using QuantConnect.AlphaStream.Infrastructure;
 using QuantConnect.AlphaStream.Models.Orders;
 using RestSharp;
+using System.Collections.Generic;
 
 namespace QuantConnect.AlphaStream.Requests
 {
@@ -9,7 +9,7 @@ namespace QuantConnect.AlphaStream.Requests
     /// Fetch the Alpha Orders and OrderEvents list(backtest and live trading) track record since publication.
     /// </summary>
     [Endpoint(Method.GET, "alpha/{id}/orders")]
-    public class GetAlphaOrdersRequest : AttributeRequest<List<Order>>
+    public class GetAlphaOrdersRequest : AttributeRequest<List<AlphaStreamOrder>>
     {
         /// <summary>
         /// Unique id hash of an Alpha published to the marketplace.
